Move ToHostname trait to traits module where it belongs
