+++ src/utils.rs
@@ -1,14 +1,14 @@
 use std::{net::IpAddr, path::Path, str::FromStr, sync::Once};
 
 use ipnetwork::IpNetwork;
-use regex::Regex;
 use tracing::warn;
-use trust_dns_resolver::IntoName;
 use trust_dns_server::client::rr::{LowerName, Name};
-use zerotier_central_api::{apis::configuration::Configuration, models::Member};
+use zerotier_central_api::apis::configuration::Configuration;
 
 use anyhow::anyhow;
 use zerotier_one_api::apis::configuration::Configuration as ZTOneConfiguration;
+
+use crate::traits::ToHostname;
 
 // collections of test hosts files
 pub const TEST_HOSTS_DIR: &str = "testdata/hosts-files";
@@ -239,74 +239,4 @@ pub async fn update_central_dns(
     }
 
     Ok(())
-}
-
-// translation_table should also be lazy_static and provides a small match set to find and correct
-// problems with member namesl.
-fn translation_table() -> Vec<(Regex, &'static str)> {
-    vec![
-        (Regex::new(r"\s+").unwrap(), "-"), // translate whitespace to `-`
-        (Regex::new(r"[^.\s\w\d-]+").unwrap(), ""), // catch-all at the end
-    ]
-}
-
-pub trait ToHostname {
-    fn to_hostname(self) -> Result<Name, anyhow::Error>;
-    fn to_fqdn(self, domain: Name) -> Result<Name, anyhow::Error>;
-}
-
-impl ToHostname for &str {
-    fn to_hostname(self) -> Result<Name, anyhow::Error> {
-        self.to_string().to_hostname()
-    }
-
-    fn to_fqdn(self, domain: Name) -> Result<Name, anyhow::Error> {
-        Ok(self.to_hostname()?.append_domain(&domain).unwrap())
-    }
-}
-
-impl ToHostname for IpAddr {
-    fn to_hostname(self) -> Result<Name, anyhow::Error> {
-        self.to_string().to_hostname()
-    }
-
-    fn to_fqdn(self, domain: Name) -> Result<Name, anyhow::Error> {
-        self.to_string().to_fqdn(domain)
-    }
-}
-
-impl ToHostname for Member {
-    fn to_hostname(self) -> Result<Name, anyhow::Error> {
-        ("zt-".to_string() + &self.node_id.unwrap()).to_hostname()
-    }
-
-    fn to_fqdn(self, domain: Name) -> Result<Name, anyhow::Error> {
-        ("zt-".to_string() + &self.node_id.unwrap()).to_fqdn(domain)
-    }
-}
-
-impl ToHostname for String {
-    // to_hostname turns member names into trust-dns compatible dns names.
-    fn to_hostname(self) -> Result<Name, anyhow::Error> {
-        let mut s = self.trim().to_string();
-        for (regex, replacement) in translation_table() {
-            s = regex.replace_all(&s, replacement).to_string();
-        }
-
-        let s = s.trim();
-
-        if s == "." || s.ends_with(".") {
-            return Err(anyhow!("Record {} not entered into catalog: '.' and records that ends in '.' are disallowed", s));
-        }
-
-        if s.len() == 0 {
-            return Err(anyhow!("translated hostname {} is an empty string", self));
-        }
-
-        Ok(s.trim().into_name()?)
-    }
-
-    fn to_fqdn(self, domain: Name) -> Result<Name, anyhow::Error> {
-        Ok(self.to_hostname()?.append_domain(&domain).unwrap())
-    }
 }
